Video Presentation
This presentation sheds light on the revolutionary potential of deep learning in the fight against lung cancer, this presentation delves into the cutting-edge models such as MobileNetV3, VGG16, ResNet50, ViT B16, and ViT B32. We research their accuracy in diagnosis by means of an engrossing story of optimism, advancement, and technological power. We shed light on the future of cancer care by emphasizing the potential of early detection and the subtleties of model performance. This discussion is about more than just statistics and facts; it's a a ray of hope, and a salute to the tenacity of cancer patients.